rag-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
langchain==0.1.16
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langchain-pinecone==0.1.0
langchain-community==0.0.32
pinecone-client==3.2.2
google-generativeai==0.3.2
pypdf==3.17.4
python-multipart==0.0.6
